Resolve NNO-7756 "Release 7.21/"

See merge request open/n2o-framework!4150

+++ backend/n2o/n2o-engine/src/main/java/net/n2oapp/framework/engine/data/json/TestDataProviderEngine.java
@@ -23,6 +23,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.chrono.ChronoLocalDate;
 import java.time.chrono.ChronoLocalDateTime;
+import java.time.format.DateTimeParseException;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
@@ -45,8 +46,14 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
      */
     private String classpathResourcePath;
 
-    private ResourceLoader resourceLoader = new DefaultResourceLoader();
+    /**
+     * Обновление данных в файле на диске
+     */
+    private boolean readonly;
+
+
     private final Map<String, List<DataSet>> repository = new ConcurrentHashMap<>();
+    private ResourceLoader resourceLoader = new DefaultResourceLoader();
     private final Map<String, AtomicLong> sequences = new ConcurrentHashMap<>();
     private ObjectMapper objectMapper;
 
@@ -75,8 +82,8 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
     protected synchronized List<DataSet> getData(N2oTestDataProvider invocation) {
         if (invocation.getFile() == null)
             return new ArrayList<>();
-        if (getRepositoryData(invocation.getFile()) == null ||
-                fileExistsOnDisk(invocation.getFile())) {
+        boolean isInit = getRepositoryData(invocation.getFile()) == null;
+        if (isInit || (!readonly && fileExistsOnDisk(invocation.getFile()))) {
             initRepository(invocation);
         }
 
@@ -130,7 +137,7 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
      * @param filename Имя файла
      */
     protected void updateFile(String filename) {
-        if (fileExistsOnDisk(filename)) {
+        if (!readonly && fileExistsOnDisk(filename)) {
             try (FileWriter fileWriter = new FileWriter(getFullPathOnDisk(filename))) {
                 String mapAsJson = objectMapper.writeValueAsString(getRepositoryData(filename));
                 fileWriter.write(mapAsJson);
@@ -489,11 +496,11 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
                             return ((Number) m.get(field)).longValue() < ((Number) pattern).longValue();
                         }
                         if (pattern instanceof LocalDate) {
-                            LocalDate date = LocalDate.parse(m.get(field).toString());
+                            LocalDate date = parseToLocalDate(m.get(field).toString());
                             return date.isEqual((ChronoLocalDate) pattern) || date.isBefore((ChronoLocalDate) pattern);
                         }
                         if (pattern instanceof LocalDateTime) {
-                            LocalDateTime dateTime = LocalDateTime.parse(m.get(field).toString());
+                            LocalDateTime dateTime = parseToLocalDateTime(m.get(field).toString());
                             return dateTime.isEqual((ChronoLocalDateTime<?>) pattern) || dateTime.isBefore((ChronoLocalDateTime<?>) pattern);
                         }
                         return m.get(field).toString().compareTo(pattern.toString()) < 0;
@@ -514,11 +521,11 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
                             return ((Long) ((Number) m.get(field)).longValue()).compareTo(((Number) pattern).longValue()) > 0;
                         }
                         if (pattern instanceof LocalDate) {
-                            LocalDate date = LocalDate.parse(m.get(field).toString());
+                            LocalDate date = parseToLocalDate(m.get(field).toString());
                             return date.isEqual((ChronoLocalDate) pattern) || date.isAfter((ChronoLocalDate) pattern);
                         }
                         if (pattern instanceof LocalDateTime) {
-                            LocalDateTime dateTime = LocalDateTime.parse(m.get(field).toString());
+                            LocalDateTime dateTime = parseToLocalDateTime(m.get(field).toString());
                             return dateTime.isEqual((ChronoLocalDateTime<?>) pattern) || dateTime.isAfter((ChronoLocalDateTime<?>) pattern);
                         }
                         return m.get(field).toString().compareTo(pattern.toString()) > 0;
@@ -645,6 +652,10 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
         this.pathOnDisk = pathOnDisk;
     }
 
+    public void setReadonly(boolean readonly) {
+        this.readonly = readonly;
+    }
+
     public String getClasspathResourcePath() {
         return classpathResourcePath;
     }
@@ -661,6 +672,14 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
         this.objectMapper = objectMapper;
     }
 
+    public boolean isReadonly() {
+        return readonly;
+    }
+
+    public Map<String, List<DataSet>> getRepository() {
+        return repository;
+    }
+
     /**
      * Проверяет существование файла на диске
      *
@@ -671,5 +690,33 @@ public class TestDataProviderEngine implements MapInvocationEngine<N2oTestDataPr
     private boolean fileExistsOnDisk(String filename) {
         return pathOnDisk != null &&
                 new File(getFullPathOnDisk(filename)).isFile();
+    }
+
+    /**
+     * Переводит строковое представление даты в LocalDate
+     * @param strDate Строковое представление даты в формате ISO_LOCAL_DATE
+     * @return        Переменную типа LocalDate, соответствующую строковому представлению даты, при неверном формате
+     * строкового представления выбрасывается N2oException
+     */
+    private LocalDate parseToLocalDate(String strDate) {
+        try {
+            return LocalDate.parse(strDate);
+        } catch (DateTimeParseException e) {
+            throw new N2oException("Формат даты, используемый в json, не соответствует ISO_LOCAL_DATE", e);
+        }
+    }
+
+    /**
+     * Переводит строковое представление даты и времени в LocalDateTime
+     * @param strDateTime Строковое представление даты и времени в формате ISO_LOCAL_DATE_TIME
+     * @return            Переменную типа LocalDateTime, соответствующую строковому представлению даты и времени, при неверном формате
+     * строкового представления выбрасывается N2oException
+     */
+    private LocalDateTime parseToLocalDateTime(String strDateTime) {
+        try {
+            return LocalDateTime.parse(strDateTime);
+        } catch (DateTimeParseException e) {
+            throw new N2oException("Формат даты и времени, используемый в json, не соответствует ISO_LOCAL_DATE_TIME", e);
+        }
     }
 }
